backend/pyproject.toml
[project]
name = "app-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "celery",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "redis",
    "pydantic-settings",
    "pyjwt",
    "passlib[argon2]",
    "python-multipart",
    "email-validator",
    "httpx",
    "pyyaml",
    "cryptography",
    "aiosmtplib",
    "pyotp",
    "boto3",
    "sentry-sdk[fastapi]",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-exporter-otlp-proto-grpc",
    "pgvector",
    "langgraph>=1.0.10,<2",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI Depends() in defaults

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501"]
"alembic/env.py" = ["E402"]
